Move get_file_diff as helper method
